support nan values within hist

+++ mlrun/common/schemas/model_monitoring/model_endpoints.py
@@ -103,7 +103,7 @@ class ModelEndpointSpec(ObjectSpec):
 
 class Histogram(BaseModel):
     buckets: List[float]
-    counts: List[float]
+    counts: List[int]
 
 
 class FeatureValues(BaseModel):
